Refactor to make the file uploaded classes less ugly with dynamically removing methods.
